% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_abundance.R
\name{mean_abundance}
\alias{mean_abundance}
\title{Calculate mean abundance of species}
\usage{
mean_abundance(data)
}
\arguments{
\item{data}{A data frame with columns for hosts (column 1) and species (columns 3 onwards).}
}
\value{
A data frame with columns for host, species, mean, sd, and se.
}
\description{
This function calculates the mean abundance of species across hosts.
}
\examples{

data <- data.frame(
  host = c("A", "A", "A", "B", "B", "B"),
  species1 = c(1, 5, 3, 2, 4, 0),
  species2 = c(3, 2, 0, 1, 0, 0)
)

mean_abundance(data)

}
