% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parNUP-package.R
\docType{package}
\name{parNUP-package}
\alias{parNUP}
\alias{parNUP-package}
\title{parNUP: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nupelia/parNUP}
  \item Report bugs at \url{https://github.com/nupelia/parNUP/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
