% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Calculate the prevalence of parasitism}
\usage{
prevalence(data)
}
\arguments{
\item{data}{A data frame with individual hosts as the first column and
parasite counts as the remaining columns}
}
\value{
A data frame with the host as the first column and prevalence as
the second
}
\description{
Calculate the prevalence of parasitism
}
\examples{
data <- data.frame(host = c("A", "A", "B", "B"),
                   parasite1 = c(1, 0, 0, 1),
                   parasite2 = c(0, 1, 0, 0))
prevalence(data)
#>   host n_parasited n_host prevalence
#> 1    A            2      2        1.0
#> 2    B            1      2        0.5

}
