% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Calculate the prevalence of parasitism}
\usage{
prevalence(data)
}
\arguments{
\item{data}{A data frame with individual hosts as the first column and parasite counts as the remaining columns}
}
\value{
A data frame with the prevalence of each species
}
\description{
Calculate the prevalence of parasitism
}
\examples{
prevalence(dataset)
}
